requirements.txt
streamlit>=1.32
pandas>=2.1
python-dotenv>=1.0
PyYAML>=6.0
